Add listV2 to support long file sizes and last modified values
list ("LIST" command) only supports int for file size/last modified.
listV2 ("LIS2" command) supports long instead.

+++ src/se/vidstige/jadb/RemoteFile.java
@@ -9,8 +9,8 @@ public class RemoteFile {
     public RemoteFile(String path) { this.path = path; }
 
     public String getName() { throw new UnsupportedOperationException(); }
-    public int getSize() { throw new UnsupportedOperationException(); }
-    public int getLastModified() { throw new UnsupportedOperationException(); }
+    public long getSize() { throw new UnsupportedOperationException(); }
+    public long getLastModified() { throw new UnsupportedOperationException(); }
     public boolean isDirectory() { throw new UnsupportedOperationException(); }
 
     public String getPath() { return path;}
