+++ src/se/vidstige/jadb/RemoteFileRecord.java
@@ -2,6 +2,8 @@ package se.vidstige.jadb;
 
 /**
  * Created by vidstige on 2014-03-19.
+ * Returned by LIST command. Only has ints for mode/size/lastModified.
+ * See {@link RemoteFileRecordV2} also.
  */
 class RemoteFileRecord extends RemoteFile {
     public static final RemoteFileRecord DONE = new RemoteFileRecord(null, 0, 0, 0);
@@ -18,12 +20,12 @@ class RemoteFileRecord extends RemoteFile {
     }
 
     @Override
-    public int getSize() {
+    public long getSize() {
         return size;
     }
 
     @Override
-    public int getLastModified() {
+    public long getLastModified() {
         return lastModified;
     }
